requirements.txt
flask
nltk
numpy
networkx
scipy
requests
pyperclip
Pillow
opencv-python
python-dotenv
